ticket & equipment request controller

+++ backend/src/main/java/com/EquipPro/backend/service/EquipmentRequestImp.java
@@ -1,6 +1,7 @@
 package com.EquipPro.backend.service;
 
 import com.EquipPro.backend.exception.EquipmentRequestNotFoundException;
+import com.EquipPro.backend.exception.UserNotFoundException;
 import com.EquipPro.backend.model.Equipment;
 import com.EquipPro.backend.model.EquipmentRequest;
 import com.EquipPro.backend.model.User;
@@ -26,22 +27,32 @@ public class EquipmentRequestImp implements EquipmentRequestService{
 
     @Override
     public EquipmentRequest getEquipmentRequest(Long equipmentRequestId) {
-        return equipmentRequestRepository.findById(equipmentRequestId)
+        EquipmentRequest equipmentRequest = equipmentRequestRepository.findById(equipmentRequestId)
                 .orElseThrow(()-> new EquipmentRequestNotFoundException("The equipment request not found!"));
+        equipmentRequest.setRequestStatus(true);
+        return equipmentRequest;
     }
 
     @Override
-    public void createEquipmentRequest(Long userId, Long equipmentId) {
+    public EquipmentRequest createEquipmentRequest(Long userId, Long equipmentId) {
         EquipmentRequest equipmentRequest = new EquipmentRequest();
         Equipment equipment = equipmentServiceImp.getEquipment(equipmentId);
         Optional<User> user = userRepository.findById(userId);
+        if (user.isEmpty()){
+            throw new UserNotFoundException("User not found");
+        }
         equipmentRequest.setEquipment(equipment);
         equipmentRequest.setWorker(user.get());
         equipmentRequest.setRequestStatus(false);
+        return equipmentRequest;
     }
 
     @Override
     public void deleteEquipmentRequest(Long equipmentRequestId) {
+        Optional<EquipmentRequest> equipmentRequest = equipmentRequestRepository.findById(equipmentRequestId);
+        if (equipmentRequest.isEmpty()){
+            throw new EquipmentRequestNotFoundException("Equipment Request not found");
+        }
         equipmentRequestRepository.deleteById(equipmentRequestId);
     }
 }
